remove unnecessary type hint

+++ problem0067_golf.hs
@@ -1,10 +1,10 @@
 -- Golfed solution for Euler #67 in Haskell.
 -- Pipe the triangle/tree file from stdin, prints result.
--- 8 lines, 171 chars without comments.
+-- 8 lines, 166 chars without comments.
 z=zipWith
 n (l:[])=maximum l
 n (x:y:l)=n$z max(z(+)x y++[0])(0:z(+)x(tail y)):l
 main=do
 c<-getContents
-let a=n$map(map read.words)$filter(notElem '#')$lines c::Int
+let a=n$map(map read.words)$filter(notElem '#')$lines c
 print a
